requirements.txt
fastapi
uvicorn
sqlalchemy
pandas
scikit-learn
python-multipart

# Auth & Security
passlib[bcrypt]
python-jose[cryptography]

# Optional but safe
python-dotenv
